Use commons-io version 2.20.0 to fix a security vulnerability introduced by org.apache.commons:commons-fileupload2-javax
The commons-fileupload2-javax has a transitive dependency on a vulnerable version of commons-io. Unfortunately,
commons-fileupload2-javax could not be further upgraded due to problems during compilation. Therefor it was decided
to upgrade its dependency on commons-io instead, in order to mitigate a potential security issue in commons-io.

In the end the whole Java 8 / Tomcat 9.0.x stack is reaching its end-of-life. It seems
that the right way forward is to migrate to newer Java and Tomcat versions.